Move AssertSnapshotTrait to Testing Bundle
