+++ packages/content/tests/Functional/Infrastructure/Sulu/Sitemap/ContentSitemapProviderTest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Sulu\Content\Tests\Functional\Infrastructure\Sulu\Sitemap;
 
+use Sulu\Bundle\TestBundle\Testing\AssertSnapshotTrait;
 use Sulu\Bundle\TestBundle\Testing\SetGetPrivatePropertyTrait;
 use Sulu\Bundle\TestBundle\Testing\SuluTestCase;
 use Sulu\Bundle\WebsiteBundle\Sitemap\Sitemap;
@@ -21,7 +22,6 @@ use Sulu\Bundle\WebsiteBundle\Sitemap\SitemapUrl;
 use Sulu\Content\Infrastructure\Sulu\Sitemap\ContentSitemapProvider;
 use Sulu\Content\Tests\Application\ExampleTestBundle\Entity\Example;
 use Sulu\Content\Tests\Application\ExampleTestBundle\Entity\ExampleDimensionContent;
-use Sulu\Content\Tests\Traits\AssertSnapshotTrait;
 use Sulu\Content\Tests\Traits\CreateExampleTrait;
 
 class ContentSitemapProviderTest extends SuluTestCase
