+++ packages/content/tests/Functional/Infrastructure/Sulu/SmartContent/ContentDataProviderTest.php
@@ -15,12 +15,12 @@ namespace Sulu\Content\Tests\Functional\Infrastructure\Sulu\SmartContent;
 
 use Sulu\Bundle\CategoryBundle\Entity\CategoryInterface;
 use Sulu\Bundle\TagBundle\Tag\TagInterface;
+use Sulu\Bundle\TestBundle\Testing\AssertSnapshotTrait;
 use Sulu\Bundle\TestBundle\Testing\SuluTestCase;
 use Sulu\Component\SmartContent\ArrayAccessItem;
 use Sulu\Component\SmartContent\DataProviderResult;
 use Sulu\Component\SmartContent\ResourceItemInterface;
 use Sulu\Content\Infrastructure\Sulu\SmartContent\Provider\ContentDataProvider;
-use Sulu\Content\Tests\Traits\AssertSnapshotTrait;
 use Sulu\Content\Tests\Traits\CreateCategoryTrait;
 use Sulu\Content\Tests\Traits\CreateExampleTrait;
 use Sulu\Content\Tests\Traits\CreateTagTrait;
